// app.client/src/api/models/InventoryItemInTransactionListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type InventoryItemInTransactionListVM = {
    inventoryItemId?: string | null;
    ownerBarcodeItemNum?: string | null;
    productDescription?: string | null;
    productNotes?: string | null;
    productUom?: string | null;
    productGrossWeightKg?: number;
    productPackage?: string | null;
    lastPricePaid?: number;
    isLotRequired?: boolean;
    isSnRequired?: boolean;
    isDateMfgRequired?: boolean;
    isDateExpRequired?: boolean;
    imageUrl?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
};